+++ apps/native-component-list/src/screens/UI/UIScreen.android.tsx
@@ -35,6 +35,14 @@ export const UIScreens = [
     },
   },
   {
+    name: 'Shape component',
+    route: 'ui/shape',
+    options: {},
+    getComponent() {
+      return optionalRequire(() => require('./ShapeScreen'));
+    },
+  },
+  {
     name: 'Section component',
     route: 'ui/section',
     options: {},